requirements.txt
google-auth==2.38.0
google-genai==1.4.0
psycopg2-binary==2.9.10
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1